tweak eg hack for less distortion in sine wave
A waveform value of '0' means shift right 16 bits. An EG output of
silence also means shift right 16 bits. This means shift right 32 bits,
which overflows the shift instruction, meaning we shift nothing and
introduce a loud pop every time we land on the 0 cycle. Obviously, use a
max instruction after arriving at the shift amount and be done with it.
Unfortunately, the timing of the instructions means we cannot do this
without actually introducing 3 instruction cycles (12 clocks), and as
the opeators are unrolled, 12 more longs into the cog's program space.
This was hacked around by using (16 << 10) - 1 as the 0 value rather
than 16 << 10, with the caveat that period of what should be silence is
actually a pop on the LSB of the audio output. This is lame, and it
slightly distorts the sine wave.

Remove the -1 value from the sine lookup table restoring sine waves back
to being undistorted and actually silent, at the expense of having this
problem for triangle waves.