use pitch envelope for pitch
Now the multipliers (which, being in pitch units, are addition) simply
add to the value of the pitch envelope. This means knowledge of which
operators are fixed frequency moves into the EG so that it doesn't bend
fixed frequencies.

This paves the way for portamento in the upcoming mono mode. The pitch
EG will be able to do all the work for that.