display sleep
rather than keep display on all the time, dim after 5 seconds, then
sleep 20 seconds after that. This should help extend the life of the
OLED display.

as the display updating is fast enough that we no longer need to delay
drawing the graph portion of the display, remove the code for making
that a time delay, gaining us room to implement the more important
display sleep.